+++ src/ionic/declarations/interfaces/State.d.ts
@@ -1,3 +1,5 @@
+/// <reference path="./Config.d.ts" />
+
 declare module "cordova-background-geolocation" {
   interface State extends Config {
     enabled: boolean;